refactor(occurrence): movendo componentes de app/ para components/
