+++ src/components/occurrences/DeleteAttachmentButton.tsx
