+++ src/components/occurrences/DeleteOccurrenceButton.tsx
